moving Location interface to its own file for backwards compatibility
