+++ packages/teams-js/test/public/geoLocation.spec.ts
@@ -3,7 +3,6 @@ import { DOMMessageEvent } from '../../src/internal/interfaces';
 import { app } from '../../src/public/app';
 import { errorNotSupportedOnPlatform, FrameContexts } from '../../src/public/constants';
 import { ErrorCode, geoLocation, location } from '../../src/public/index';
-import { Location } from '../../src/public/interfaces';
 import { _minRuntimeConfigToUninitialize } from '../../src/public/runtime';
 import { FramelessPostMocks } from '../framelessPostMocks';
 
@@ -18,7 +17,7 @@ describe('geoLocation', () => {
     allowChooseLocation: true,
     showMap: true,
   };
-  const defaultLocation: Location = { latitude: 17, longitude: 17, accuracy: -1, timestamp: 100 };
+  const defaultLocation: geoLocation.Location = { latitude: 17, longitude: 17, accuracy: -1, timestamp: 100 };
   const originalDefaultPlatformVersion = '1.6.0';
 
   beforeEach(() => {
